Add a new section, Joy to Balloon entry
This commit adds a new section to the Balloon entry, talking about the joy of balloons and including a summary of the Joy article. Additionally, the Selflessness section has been edited to better reflect the situation. The Time off 2022-12 commit has also been modified, with a new item of 'Doing programming with Zach'.