Replace Writing Image and Add Floating Right Image
This commit replaces a previous image of the pen with a new image of a raccoon with a quill, and it also adds a new floating right image element. The updated image better illustrates the content of the post, and the new floating right image serves to spruce up the post and draw attention.