Update Science and Machine Learning Sections
Updating the Science and Machine Learning sections of the Techdiary with new content on Natural Language Processing, Computer Vision and Recommenders and Ranking, to better reflect the current breadth of topics in tech. In the Science section I've added a Natural Language Processing subsection, in Machine Learning I've added subsections on Natural Language Processing, Computer Vision and Recommenders and Ranking. In both sections I've also updated content on Recommendation and GPT-3.