Removing unused assets and updating asciinema syntax
In this commit, we removed the asciinema-player JS script and CSS stylesheet as they were not being used. To update asciinema syntax, we replaced Script 1 with the asciinema-player source code, while Script 2 and Script 3 were updated to include the source code, and a Using Player Syntax block was added to list and explain the updated syntax. Finally, the text before and after the player was updated and clarified.