Add page.imagefeatureblob to DRY blob image path
This commit addresses an adjustment to the page.imagefeature variable in the _includes/head.html template. For pages that now have the page.imagefeatureblob variable, this commit uses that instead of the previous page.imagefeature variable. This commit ensures that pages with the page.imagefeatureblob variable are properly rendered in the template.