Update Physical Health Post with Kettle Bells and Books
This commit updates the post about physical health with a new section about Kettle Bells and recommended books to better understand how body parts get hurt and how to address it. Kettle bell exercises are demonstrated with a video, and the books described with link to their Amazon pages. The post has also been updated to correct spelling errors.