Refactor Performance Management Section
Refactored the performance management section of the post, to more accurately reflect current best practices in the industry. Updated terminology from "firing" to "managing out," as well as renamed section headings for better accuracy. Revised the "players" involved in the process, as well as updated language in the relevant sections. Added a new section on how to handle poor communication, ESL, and ineffective communication in some modalities.