Update Docker run commands in package.json
This commit updates the Docker run commands in the package.json file to address a discrepancy between the local and container environments. Previously, running the 'docker:run' command would fail to properly serve the Jekyll site within the Docker container.

To resolve this issue, the 'docker:run' command has been modified to explicitly call the 'jekyll:container' npm script, ensuring that the Jekyll server runs correctly within the Docker container. Additionally, this command now maps both the livereload and server ports (35729 and 4000) to their corresponding ports in the Docker container.

A new 'docker:run2' command has also been added for situations where only the volume needs to be mounted within the Docker container, without running the server or mapping ports.

These changes improve the flexibility and reliability of the Docker workflow for this project.